requirements.txt
flask==3.0.3
flask-login==0.6.3
pymongo==4.8.0
python-dotenv==1.0.1
pdfminer.six==20231228
sentence-transformers==3.0.1
scikit-learn==1.5.1
groq==0.9.0
httpx==0.27.2 
h11==0.14.0
httpcore==1.0.5